Create reusable mathml2png library from mathmltable2png
Pull the reusable parts out into a library
